Allows fractional shield size values
Updates the shield size to accept double values for finer control over the particle removal distance.

This change improves the configuration options and provides a more precise user experience.

+++ src/main/java/wvv/particleshield/commands/CommandShieldSize.java
@@ -27,7 +27,7 @@ public class CommandShieldSize extends CommandBase {
         }
 
         try {
-            int newSize = Integer.parseInt(args[0]);
+            double newSize = Double.parseDouble(args[0]);
             if (newSize < 0) {
                 newSize = 0;
             }
